Started work on using socket-io to update the user's trace flags.

+++ app/components/trace-flag.js
@@ -1,10 +1,12 @@
 import Ember from 'ember';
+import config from 'sith/config/environment';
 
 export default Ember.Component.extend({
 
     store: Ember.inject.service(),
     user: Ember.inject.service(),
     notify: Ember.inject.service(),
+    io: Ember.inject.service('socket-io'),
 
     isEditingNewTraceFlag: false,
 
@@ -14,12 +16,33 @@ export default Ember.Component.extend({
     //Will hold the debug level chosen by the user in the create trace flag form.
     newDebugLevel: null,
 
-    didReceiveAttrs() {
-        let debugLevels = this.get('store').peekAll('debug-level');
-        this.set('debugLevels', debugLevels);
+    //Used to determine if a test run is executing when this trace flag component is rendered. Used to
+    //prevent the user from manipulating their traceflags during the test run which could affect analytics collecting.
+    asyncApexJobs: null,
+
+    activeAsyncApexJobs: Ember.computed.filterBy('asyncApexJobs', 'status', 'Processing'),
+
+    //Used to disabled the ability to manipulate trace flags during a test run.
+    testsExecuting: Ember.computed.notEmpty('activeAsyncApexJobs'),
+
+    init() {
+
+        //Make sure Ember is able to do important setup work (if any);
+        this._super(...arguments);
+
+        let io = this.get('io').socketFor(`${config.APP.apiDomain}/`);
+        io.on('sobject-update', this.sobjectUpdate, this);
     },
 
-    //Fired after every render/re-render of this component.
+    //DEFAULT EMBER HOOK
+    didReceiveAttrs() {
+        let store = this.get('store');
+        this.set('debugLevels', store.peekAll('debug-level'));
+        this.set('asyncApexJobs', store.peekAll('async-apex-job'));
+    },
+
+    //Fired after every render/re-render of this component, primarily used for DOM manipulation.
+    //DEFAULT EMBER HOOK
     didRender() {
         this.$('[data-toggle="tooltip"]').tooltip({
             delay: {
@@ -30,8 +53,25 @@ export default Ember.Component.extend({
     },
 
     //Fired when this component is about to be removed from the DOM.
+    //DEFAULT EMBER HOOK
     willDestroyElement() {
         this.$('[data-toggle="tooltip"]').tooltip('destroy');
+    },
+
+    sobjectUpdate(message) {
+
+        let store = this.get('store');
+        let notify = this.get('notify');
+
+        let sobjectType = message.type.dasherize();
+
+        if(message.action === 'delete') {
+            let sobject = store.peekRecord(sobjectType, message.id);
+            if(sobject) {
+                store.unloadRecord(sobject);
+                notify.success('trace flag was deleted in the current org');
+            }
+        }
     },
 
     actions: {
@@ -39,6 +79,11 @@ export default Ember.Component.extend({
         updateTraceFlag(traceFlag, hours) {
 
             let notify = this.get('notify');
+
+            if(this.get('testsExecuting')) {
+                notify.error('cannot update trace flags during test execution');
+                return;
+            }
 
             traceFlag.setProperties({
                 'startDate': moment().format(),
@@ -50,12 +95,12 @@ export default Ember.Component.extend({
                 console.info('updating trace flag result =>', result);
                 notify.success('trace flag updated');
 
-            }).catch(error => {
+            }).catch(ex => {
 
                 traceFlag.rollbackAttributes();
-                notify.error('failed to update trace flag');
-                console.error('failed to update trace flag =>', error);
 
+                ex.errors.forEach(error => notify.error({ html: `<h4>${error.title}</h4><div>${error.detail}</div>` }));
+                console.error('failed to update trace flag =>', ex);
             });
         },
 
@@ -75,6 +120,12 @@ export default Ember.Component.extend({
         saveNewTraceFlag(hours) {
 
             let notify = this.get('notify');
+
+            if(this.get('testsExecuting')) {
+                notify.error('cannot create trace flags during test execution');
+                return;
+            }
+
             let debugLevel = this.get('newDebugLevel');
 
             if(!debugLevel) {
@@ -121,6 +172,11 @@ export default Ember.Component.extend({
 
             let notify = this.get('notify');
 
+            if(this.get('testsExecuting') && traceFlag.get('debugLevel.developerName') === 'APEX_ANALYTICS') {
+                notify.error('cannot delete the APEX_ANALYTICS trace flag while tests are running');
+                return;
+            }
+
             //Will send a delete request to the server.
             traceFlag.destroyRecord().then(result => {
                 console.info('delete result', result);
@@ -141,12 +197,7 @@ export default Ember.Component.extend({
             let reloadPromise = this.get('reload')();
 
             reloadPromise.then(traceFlags => {
-
-                let count = traceFlags.get('length');
-
-                notify.success('trace flags refreshed');
-                notify.info(`${count} trace flag${count !== 1 ? 's' : ''} found`);
-
+                notify.success(`trace flags refreshed, found ${traceFlags.get('length')}`);
             }).catch(error => {
                 notify.error(error.message);
             }).finally(() => {
